restore the test to test_json and replace pandas by Z class
